refactor(tests): extract BoxCleanup RAII guard to test-utils
The detached-box SIGKILL cleanup struct landed inline in
`src/cli/tests/dind_port_conflict.rs` in commit 9333e2a (65 lines:
struct + impl Drop scanning /proc/*/fd to find live libkrun VMs
holding files under <home>/boxes/<box_id>/). A follow-up integration
test (`port_publish_host_reachability`) needs the same teardown for
the same reason — `boxlite run -d` daemonizes and removes the
on-disk shim.pid before Drop time, and `boxlite rm -f` can't be
trusted because rt_impl's recovery path mis-identifies the live
shim as dead (the May 2026 incident).

Moves the type verbatim to `src/test-utils/src/box_cleanup.rs` and
re-exports it as `boxlite_test_utils::box_cleanup::BoxCleanup`. The
implementation is unchanged; the module gains a 27-line doc block
that pins the two non-obvious design decisions:
  - Scan /proc/*/fd, don't read shim.pid (race with daemonize).
  - Don't use `boxlite rm -f` (false-negative recovery path).
  - Declare AFTER PerTestBoxHome so reverse-order Drop tears down
    the box BEFORE the home dir is removed (matching needle would
    no longer match otherwise → kill silently no-ops → leak).

`dind_port_conflict.rs` swaps its inline definition for the import
(-65 / +1). The unused `use std::path::PathBuf;` drops out with it.
